Redesign error handling for decoding.
This commit introduces an Error trait for decoding errors and switches
the Source trait to having to provide an error type implementing this
trait. Users have to call methods of this trait when creating an error
which will allow them to add additional error information.

While at it, the name of the associated type in Source was changed from
Err to Error in line with common practice.

This is a breaking change.